chore: move and update colorhash
